refactor: move utility.py to helpers-dir
