+++ pages/neem_overview.py
@@ -7,7 +7,7 @@ from threading import Lock
 from html_sanitizer import Sanitizer
 from html_sanitizer.sanitizer import sanitize_href, bold_span_to_strong,italic_span_to_em, target_blank_noopener, tag_replacer
 
-from utility import copy_dir, copy_file, download_file, make_archive_of_files_and_dirs, mutex_lock, read_file, remove_if_is_file, start_thread, unzip_file, remove_if_is_dir, write_non_binary_file, dump_dict_to_json, get_dict_from_json
+from helpers.utility import copy_dir, copy_file, download_file, make_archive_of_files_and_dirs, mutex_lock, read_file, remove_if_is_file, start_thread, unzip_file, remove_if_is_dir, write_non_binary_file, dump_dict_to_json, get_dict_from_json
 from config.settings import WEBROB_PATH, STATIC_DIR_PATH, DEFAULT_FILES_PATH, DOWNLOADS_DIR_PATH
 from neems.neemhub import instance as neemhub, NEEMHubConnectionError
 from neems.neem import DEFAULT_IMAGE_PATH, DEFAULT_IMAGE_PATH_NO_STATIC
