+++ runserver.py
@@ -19,7 +19,7 @@ from wtforms.validators import ValidationError
 from pathlib2 import Path
 
 from app_and_db import app, db
-from utility import random_string, oe_password_validator
+from helpers.utility import random_string, oe_password_validator
 from helpers.background_scheduler import start_background_scheduler
 from pages.publications import load_default_publications_and_papers, download_and_update_papers_and_bibtex
 from pages.neem_overview import download_neem_files, load_default_overview_files
